requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.32.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-server==1.23.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==2.0.4
multidict==6.0.4
mysql-connector-python==8.1.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.5
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.3
openai==0.27.8
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.15.0
prompt-toolkit==3.0.29
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.31.0
rich==13.4.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.25.0
tenacity==8.2.2
terminado==0.17.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.65.0
traitlets==5.5.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
yarl==1.9.2
zipp==3.10.0
